+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # [develop](https://github.com/adhearsion/adhearsion-ivr)
 
+# [v0.2.0](https://github.com/adhearsion/adhearsion-ivr/compare/0.1.0...0.2.0) - [2015-03-27](https://rubygems.org/gems/adhearsion-ivr/versions/0.2.0)
+  * Feature: Added support for defining alternative prompt resolution algorithms by overriding `#prompts`
+  * Feature: Added support for overriding `#max_attempts`
+  * Feature: Added support for specifying a timeout using #timeout
+  * Feature: Added support for specifying a grammar URL using #grammar_url
+  * Feature: Added support for specifying a validation callback
+  * Feature: Added support for specifying a `renderer`
+
 # [v0.1.0](https://github.com/adhearsion/adhearsion-ivr/compare/2c7ff73f5d6471be23e291c7d6c7b61d0128e09a...0.1.0) - [2014-03-12](https://rubygems.org/gems/adhearsion-ivr/versions/0.1.0)
   * Initial release